Resume: Refactored Code
