+++ Jobs4SMCYouth/app/src/main/java/com/example/jobs4smcyouth/MainActivity.java
@@ -13,11 +13,11 @@ import android.view.MenuItem;
 import android.widget.TextView;
 
 import com.example.jobs4smcyouth.Fragments.AboutFragment;
-import com.example.jobs4smcyouth.Fragments.ApplicationRulesFragment;
-import com.example.jobs4smcyouth.Fragments.ApplicationFragment;
+import com.example.jobs4smcyouth.Fragments.Application.ApplicationRulesFragment;
+import com.example.jobs4smcyouth.Fragments.Application.ApplicationFragment;
 import com.example.jobs4smcyouth.Fragments.InterviewFragment;
 import com.example.jobs4smcyouth.Fragments.JobPostFragment;
-import com.example.jobs4smcyouth.Fragments.JobSiteFragment;
+import com.example.jobs4smcyouth.Fragments.JobSites.JobSiteFragment;
 import com.example.jobs4smcyouth.Fragments.RequirementFragment;
 import com.example.jobs4smcyouth.Fragments.Resume.ResumeFragment;
 import com.example.jobs4smcyouth.Fragments.ScholarshipFragment;
