+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_application.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="15dp"
-    tools:context=".Fragments.ApplicationFragment">
+    tools:context=".Fragments.Application.ApplicationFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/applicationFrag_recyclerView"
