+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_job_site.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Fragments.JobSiteFragment">
+    tools:context=".Fragments.JobSites.JobSiteFragment">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/jobListFrag_recyclerView"
