+++ Jobs4SMCYouth/app/src/main/res/layout/fragment_resume_websites.xml
@@ -10,35 +10,4 @@
 
     </android.support.v7.widget.RecyclerView>
 
-
-    <TextView
-        android:id="@+id/resumeFrag_websites_TextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:text="Websites: "/>
-
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Careeronestop.org Damngood.com    Eresumes.com"/>
-
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Resume-help.org   Resumehelp.com  Myfuture.com"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Susanireland.com  Truecareers.com Monster.com"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Quintcareers.com/resres.html"/>
-
-
 </LinearLayout>
